Вносит правки согласно ревью от 13.03.2021
